make the notebook compatible with papermill
Papermill (https://papermill.readthedocs.io/en/latest/index.html)
is a tool for parameterizing and executing notebooks.

The main requirement is that all parameters must be in a single
cell with the tag `parameters` set for the cell. I also provided
type hints so that papermill can display that information to users
(I used old-style comment type hints so they would be less distracting
to those who aren't used to them.)